[refactor] #186 - 닉네임 편집 뷰,뷰컨트롤러 분리 (NicknameEditView)
