+++ Carmu/Sources/Controllers/MyPage/MyPageViewController.swift
@@ -41,7 +41,6 @@ final class MyPageViewController: UIViewController {
             target: self,
             action: #selector(showSettings)
         )
-        navigationController?.view.addSubview(myPageView.darkOverlayView)
 
         // 유저 정보 불러와서 닉네임, 프로필 표시
         guard let databasePath = User.databasePathWithUID else {
@@ -64,18 +63,6 @@ final class MyPageViewController: UIViewController {
         // 버튼에 Target 추가
         myPageView.nicknameEditButton.addTarget(self, action: #selector(showTextField), for: .touchUpInside)
         myPageView.profileImageEditButton.addTarget(self, action: #selector(showProfileChangeView), for: .touchUpInside)
-        myPageView.textFieldEditCancelButton.addTarget(self, action: #selector(dismissTextField), for: .touchUpInside)
-        myPageView.textFieldEditDoneButton.addTarget(self, action: #selector(changeNickname), for: .touchUpInside)
-
-        // darkOverlayView가 내비게이션 바까지 덮을 수 있도록 내비게이션 컨트롤러의 뷰에 대해 subView로 추가해준다.
-        navigationController?.view.addSubview(myPageView.darkOverlayView)
-        myPageView.darkOverlayView.snp.makeConstraints { make in
-            make.edges.equalToSuperview()
-        }
-        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(dismissTextField))
-        myPageView.darkOverlayView.addGestureRecognizer(tapGesture)
-
-        myPageView.nicknameTextField.delegate = self
     }
 
     override func viewWillAppear(_ animated: Bool) {
@@ -99,31 +86,11 @@ extension MyPageViewController {
         navigationController?.pushViewController(settingsVC, animated: true)
     }
 
-    // 어두운 뷰 탭하면 텍스트 필드를 포함하고 있는 darkOverlayView 비활성화
-    @objc private func dismissTextField() {
-        myPageView.darkOverlayView.isHidden = true
-        myPageView.nicknameTextField.resignFirstResponder()
-    }
-
     // 닉네임 편집 버튼을 누르면 텍스트 필드를 포함하고 있는 darkOverlayView 활성화
     @objc private func showTextField() {
-        myPageView.nicknameTextField.text = myPageView.nicknameLabel.text
-        myPageView.darkOverlayView.isHidden = false
-        myPageView.nicknameTextField.becomeFirstResponder()
-    }
-
-    // [확인] 혹은 키보드의 엔터 버튼을 눌렀을 때 닉네임 수정사항을 DB에 반영해주는 메서드
-    @objc private func changeNickname() {
-        guard let databasePath = User.databasePathWithUID else {
-            return
-        }
-        guard let newNickname = myPageView.nicknameTextField.text else {
-            return
-        }
-        databasePath.child("nickname").setValue(newNickname as NSString)
-
-        myPageView.nicknameLabel.text = myPageView.nicknameTextField.text
-        dismissTextField()
+        let nicknameEditVC = NicknameEditViewController(nickname: myPageView.nicknameLabel.text ?? "")
+        nicknameEditVC.modalPresentationStyle = .overCurrentContext
+        present(nicknameEditVC, animated: false)
     }
 
     // 프로필 설정 버튼 클릭 시 호출
@@ -133,15 +100,6 @@ extension MyPageViewController {
         profileChangeVC.selectedProfileImageColorIdx = selectedProfileImageColorIdx
         profileChangeVC.modalPresentationStyle = .formSheet
         present(profileChangeVC, animated: true)
-    }
-}
-
-// MARK: - 텍스트 필드 델리게이트 구현
-extension MyPageViewController: UITextFieldDelegate {
-
-    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
-        changeNickname()
-        return true
     }
 }
