+++ Carmu/Sources/Utils/CustomVisualEffectView.swift
@@ -1,0 +1,45 @@
+//
+//  CustomVisualEffectView.swift
+//  Carmu
+//
+//  Created by 김영빈 on 2023/10/11.
+//
+
+import UIKit
+
+// MARK: - 커스텀으로 강도를 설정해서 Visual Effect를 추가하기 위한 뷰
+final class CustomVisualEffectView: UIVisualEffectView {
+
+    /// Create visual effect view with given effect and its intensity
+    ///
+    /// - Parameters:
+    ///   - effect: visual effect, eg UIBlurEffect(style: .dark)
+    ///   - intensity: custom intensity from 0.0 (no effect) to 1.0 (full effect) using linear scale
+    init(effect: UIVisualEffect, intensity: CGFloat) {
+        theEffect = effect
+        customIntensity = intensity
+        super.init(effect: nil)
+    }
+
+    required init?(coder aDecoder: NSCoder) {
+        nil
+    }
+
+    deinit {
+        animator?.stopAnimation(true)
+    }
+
+    override func draw(_ rect: CGRect) {
+        super.draw(rect)
+        effect = nil
+        animator?.stopAnimation(true)
+        animator = UIViewPropertyAnimator(duration: 1, curve: .linear) { [unowned self] in
+            self.effect = theEffect
+        }
+        animator?.fractionComplete = customIntensity
+    }
+
+    private let theEffect: UIVisualEffect
+    private let customIntensity: CGFloat
+    private var animator: UIViewPropertyAnimator?
+}
